+++ shop-service/src/main/java/com/shoptemplate/service/UserService.java
@@ -9,7 +9,7 @@ public interface UserService {
 
     List<User> getAllUsers();
 
-    Optional<User> getUserById(int id);
+    Optional<User> getUserById(Long id);
 
     void createUser(User user);
